backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sentence-transformers
faiss-cpu       
pdfplumber
numpy
openai
python-dotenv
requests
